requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
openai-whisper==20231117
ffmpeg-python==0.2.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.18
Pillow
nest-asyncio
